Remove localstorage from Tyr
