+++ Tyr/app/services/i18n.ts
@@ -4,8 +4,8 @@ import { TranslationJson } from 'i18n-proto';
 // TODO: exclude from primary bundle
 import langRu from '../i18n/ru.json';
 import langDe from '../i18n/de.json';
-import { IDB } from './idb';
 import { environment } from '#config';
+import { IStorage } from '#/services/storage';
 export const supportedLanguages = ['en', 'ru', 'de'];
 const langs = {
   ru: langRu,
@@ -16,7 +16,7 @@ export class I18nService {
   protected i18nController: TranslationController;
   protected i18n: TranslationProvider;
 
-  constructor(private readonly storage: IDB) {
+  constructor(private readonly storage: IStorage) {
     this.i18nController = new TranslationController(
       /* translationGetter: */ (
         name: string,
